fix rake file

+++ Rakefile
@@ -1,9 +1,11 @@
 require File.expand_path('../config/application', __FILE__)
 Rails.application.load_tasks
 
-Rake::Task["test:run"].enhance ["test:features"]
+if Rails.env.production? == false
+  Rake::Task["test:run"].enhance ["test:features"]
 
-Rails::TestTask.new("test:features" => "test:prepare") do |t|
-  Rake::Task['db:reset'].invoke
-  t.pattern = "test/features/**/*_test.rb"
+  Rails::TestTask.new("test:features" => "test:prepare") do |t|
+    Rake::Task['db:reset'].invoke
+    t.pattern = "test/features/**/*_test.rb"
+  end
 end
